+++ src/components/SharePopover/__stories__/Showcase/index.tsx
@@ -4,22 +4,21 @@ import {block} from '../../../utils/cn';
 
 import {Icon} from '../../../Icon';
 import {Button} from '../../../Button';
-import {ShareTooltip, ShareOptions, ShareList} from '../../../ShareTooltip';
+import {SharePopover, ShareOptions, ShareList} from '../../../SharePopover';
 
 import {LayoutDirection} from '../../constants';
 import {Custom, Cloud, ShareArrowUp} from './icons';
-import {SocialShareData} from 'src/components/ShareTooltip/models';
+import {ShareOptionsData} from 'src/components/SharePopover/models';
 
-import './ShareTooltip.scss';
+import './SharePopover.scss';
 
-const b = block('share-tooltip-demo');
+const b = block('share-popover-demo');
 
-export function ShareTooltipDemo() {
-    const url =
-        'https://preview.yandexcloud.dev/uikit/?path=/story/components-sharetooltip--default';
+export function SharePopoverDemo() {
+    const url = 'https://www.example.com';
     const title = 'Check out this new awesome sharing component';
-    const text = 'Content sharing is not supported in all social networks';
-    const socialNets = [
+    const text = 'Content sharing is not supported in all share options';
+    const shareOptions = [
         ShareOptions.Telegram,
         ShareOptions.Facebook,
         ShareOptions.Twitter,
@@ -42,27 +41,27 @@ export function ShareTooltipDemo() {
         >
             <div style={{margin: 16, display: 'flex', alignItems: 'center'}}>
                 <span style={{marginRight: 8}}>Copy only</span>
-                <ShareTooltip url={url} title={title} text={text} />
+                <SharePopover url={url} title={title} text={text} />
             </div>
 
             <div style={{margin: 16, display: 'flex', alignItems: 'center'}}>
-                <span style={{marginRight: 8}}>Socials only</span>
-                <ShareTooltip
+                <span style={{marginRight: 8}}>Default share options only</span>
+                <SharePopover
                     url={url}
                     title={title}
                     text={text}
-                    socialNets={socialNets}
+                    shareOptions={shareOptions}
                     withCopyLink={false}
                 />
             </div>
 
             <div style={{margin: 16, display: 'flex', alignItems: 'center'}}>
-                <span style={{marginRight: 8}}>Custom Social</span>
-                <ShareTooltip
+                <span style={{marginRight: 8}}>Custom share option</span>
+                <SharePopover
                     url={url}
                     title={title}
                     text={text}
-                    socialNets={socialNets}
+                    shareOptions={shareOptions}
                     withCopyLink={false}
                     openByHover={false}
                 >
@@ -70,18 +69,18 @@ export function ShareTooltipDemo() {
                         icon={Custom}
                         url="mailto:example@example.com"
                         label="Custom"
-                        getShareLink={(params: SocialShareData) => params.url}
+                        getShareLink={(params: ShareOptionsData) => params.url}
                     />
-                </ShareTooltip>
+                </SharePopover>
             </div>
 
             <div style={{margin: 16, display: 'flex', alignItems: 'center'}}>
-                <span style={{marginRight: 8}}>Custom Social and Copy Link</span>
-                <ShareTooltip
+                <span style={{marginRight: 8}}>Custom share option and Copy Link</span>
+                <SharePopover
                     url={url}
                     title={title}
                     text={text}
-                    socialNets={socialNets}
+                    shareOptions={shareOptions}
                     withCopyLink={true}
                     openByHover={false}
                 >
@@ -89,72 +88,72 @@ export function ShareTooltipDemo() {
                         icon={Custom}
                         url="mailto:example@example.com"
                         label="Custom"
-                        getShareLink={(params: SocialShareData) => params.url}
+                        getShareLink={(params: ShareOptionsData) => params.url}
                     />
-                </ShareTooltip>
+                </SharePopover>
             </div>
 
             <div style={{margin: 16, display: 'flex', alignItems: 'center'}}>
-                <span style={{marginRight: 8}}>Socials with copy</span>
-                <ShareTooltip url={url} title={title} text={text} socialNets={socialNets} />
+                <span style={{marginRight: 8}}>Default share options with copy</span>
+                <SharePopover url={url} title={title} text={text} shareOptions={shareOptions} />
             </div>
 
             <div style={{margin: 16, display: 'flex', alignItems: 'center'}}>
-                <span style={{marginRight: 8}}>Socials with copy (only URL)</span>
-                <ShareTooltip url={url} socialNets={socialNets} />
+                <span style={{marginRight: 8}}>Share options with copy (only URL)</span>
+                <SharePopover url={url} shareOptions={shareOptions} />
             </div>
 
             <div style={{margin: 16, display: 'flex', alignItems: 'center'}}>
-                <span style={{marginRight: 8}}>Socials with copy (open by click)</span>
-                <ShareTooltip
+                <span style={{marginRight: 8}}>Share options with copy (open by click)</span>
+                <SharePopover
                     url={url}
                     title={title}
                     text={text}
-                    socialNets={socialNets}
+                    shareOptions={shareOptions}
                     openByHover={false}
                 />
             </div>
 
             <div style={{margin: 16, display: 'flex', alignItems: 'center'}}>
                 <span style={{marginRight: 8}}>Using Web Share API (for mobile)</span>
-                <ShareTooltip
+                <SharePopover
                     url={url}
                     title={title}
                     text={text}
-                    socialNets={socialNets}
+                    shareOptions={shareOptions}
                     useWebShareApi={true}
                 />
             </div>
 
             <div style={{margin: 16, display: 'flex', alignItems: 'center'}}>
                 <span style={{marginRight: 8}}>Layout with direction = &quot;column&quot;</span>
-                <ShareTooltip
+                <SharePopover
                     url={url}
                     title={title}
                     text={text}
-                    socialNets={socialNets}
+                    shareOptions={shareOptions}
                     direction={LayoutDirection.Column}
                 />
             </div>
 
             <div style={{margin: 16, display: 'flex', alignItems: 'center'}}>
                 <span style={{marginRight: 8}}>With alternative Icon</span>
-                <ShareTooltip
+                <SharePopover
                     url={url}
                     title={title}
                     text={text}
-                    socialNets={socialNets}
+                    shareOptions={shareOptions}
                     customIcon={ShareArrowUp}
                 />
             </div>
 
             <div style={{margin: 16, display: 'flex', alignItems: 'center'}}>
                 <span style={{marginRight: 8}}>With alternative Icon and Title on the right</span>
-                <ShareTooltip
+                <SharePopover
                     url={url}
                     title={title}
                     text={text}
-                    socialNets={socialNets}
+                    shareOptions={shareOptions}
                     customIcon={ShareArrowUp}
                     buttonTitle={ShareTitle}
                 />
@@ -164,11 +163,11 @@ export function ShareTooltipDemo() {
                 <span style={{marginRight: 8}}>
                     With alternative Icon and Title on the right and custom style
                 </span>
-                <ShareTooltip
+                <SharePopover
                     url={url}
                     title={title}
                     text={text}
-                    socialNets={socialNets}
+                    shareOptions={shareOptions}
                     customIcon={ShareArrowUp}
                     buttonTitle={ShareTitle}
                     switcherClassName={b('custom-switcher')}
@@ -182,11 +181,11 @@ export function ShareTooltipDemo() {
 
             <div style={{margin: 16, display: 'flex', alignItems: 'center'}}>
                 <span style={{marginRight: 8}}>Custom copy button render</span>
-                <ShareTooltip
+                <SharePopover
                     url={url}
                     title={title}
                     text={text}
-                    socialNets={socialNets}
+                    shareOptions={shareOptions}
                     openByHover={false}
                     copyTitle={ShareTitle}
                     copyIcon={Cloud}