+++ src/components/SharePopover/index.ts
@@ -1,0 +1,6 @@
+export {ShareOptions} from './constants';
+export {ShareList} from './ShareList/ShareList';
+export {SharePopover} from './SharePopover';
+
+export type {ShareListProps} from './ShareList/ShareList';
+export type {SharePopoverProps} from './SharePopover';