+++ src/components/ShareTooltip/models.ts
@@ -1,4 +1,0 @@
-export interface SocialShareData extends ShareData {
-    /** sharing link */
-    url: string;
-}